Ensure a valid ID is passed when using guest checkout.
When using guest checkout, IT_Exchange_Customer::$id is an email address. When
passing the customer ID to Auth.net the $id property was always used. When the
email was greater than 20 characters, Auth.net would reject the request. Now,
if the id is an email address, we hash it and use the first 20 characters as
the customer ID in Auth.net.